+++ src/store/base/reducer.js
@@ -1,11 +1,13 @@
 // Base reducer
 
 import initialState from '../initialState';
+
 import {
     LOAD_PLACEHOLDER_SUCCESS,
     LOAD_HEADER_IMAGE_SUCCESS,
     LOAD_SOCIALS_SUCCESS,
-    LOAD_ABOUT_ITEMS_SUCCESS
+    LOAD_ABOUT_ITEMS_SUCCESS,
+    LOAD_FOOTER_BLURB_SUCCESS
 } from './types';
 
 // Base app reducer
@@ -38,6 +40,13 @@ export default function baseReducer(state = initialState.base, action) {
             return {
                 ...state,
                 aboutItems: action.aboutItems
+            }
+
+        // Successful footer blurb load type
+        case LOAD_FOOTER_BLURB_SUCCESS:
+            return {
+                ...state,
+                footerBlurb: action.footerBlurb
             }
 
         // Default case