+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [ScrollReveal.js](https://scrollrevealjs.org) to project dependecies
 - Added scroll reveal to project home section
 - Switched header image thunks from blog [Contentful](https://www.contentful.com/) client to website content [Contentful](https://www.contentful.com/) client
+- Connected footer blurb to website content [Contentful](https://www.contentful.com/) client
 
 #### Removed